:recycle: 컬럼 셀렉터 컴포넌트 분리
