+++ src/components/kanban-task/index.ts
@@ -1,5 +1,6 @@
 export * from './task';
 export * from './task-add-dialog';
-export * from './task-edit-view-dialog';
-export * from './task-edit-form-content';
-export * from './task-view-content';
+export * from './task-detail-dialog';
+export * from './task-edit-form';
+export * from './task-details';
+export * from './task-column-selector';
